# Conflicts:
#	test/Catch2

conflict resolution: manually update Catch2 submodule to tag v3.5.0
(same as in develop branch).

+++ include/ISgfcMessage.h
@@ -1,5 +1,5 @@
 // -----------------------------------------------------------------------------
-// Copyright 2020 Patrick Näf (herzbube@herzbube.ch)
+// Copyright 2024 Patrick Näf (herzbube@herzbube.ch)
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -76,13 +76,13 @@ namespace LibSgfcPlusPlus
     /// that caused the message. Returns SgfcConstants::InvalidLineNumber if
     /// the message does not refer to a specific line in the SGF data (most
     /// fatal errors, but also some errors and warnings).
-    virtual int GetLineNumber() const = 0;
+    virtual unsigned long GetLineNumber() const = 0;
 
     /// @brief Returns the number of the column (1-based) in the parsed SGF data
     /// that caused the message. Returns SgfcConstants::InvalidColumnNumber if
     /// the message does not refer to a specific column in the SGF data (most
     /// fatal errors, but also some errors and warnings).
-    virtual int GetColumnNumber() const = 0;
+    virtual unsigned long GetColumnNumber() const = 0;
 
     /// @brief Returns true if the message describes a critical problem. Only
     /// warning and error messages can be critical. Always returns false for
