+++ src/sgfc/message/SgfcMessageStream.cpp
@@ -1,5 +1,5 @@
 // -----------------------------------------------------------------------------
-// Copyright 2020 Patrick Näf (herzbube@herzbube.ch)
+// Copyright 2024 Patrick Näf (herzbube@herzbube.ch)
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -87,8 +87,8 @@ namespace LibSgfcPlusPlus
     // there are no thousands separators
     formattedMessageTextStream.imbue(std::locale::classic());
 
-    int lineNumber = sgfcError->row;
-    int columnNumber = sgfcError->col;
+    unsigned long lineNumber = sgfcError->row;
+    unsigned long columnNumber = sgfcError->col;
     if (lineNumber > 0 && columnNumber > 0)
     {
       formattedMessageTextStream
